study_buddy_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
openai
google-api-python-client
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
python-multipart